% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{bibentry2text}
\alias{bibentry2text}
\title{Slight Improvement over \code{utils:::format.bibentry}}
\usage{
bibentry2text(x)
}
\arguments{
\item{x}{a \link[utils]{citation} and/or \link[utils]{bibentry}}
}
\description{
Slight Improvement over \code{utils:::format.bibentry}
}
\details{
Function \code{\link[=bibentry2text]{bibentry2text()}} beautifies the output from
function \code{utils:::format.bibentry(., style = 'text')}
in the following ways.
\itemize{
\item{Line break \code{'\\n'} is replaced by a white space;}
\item{Fancy quotes \eqn{``}, \eqn{''}, \eqn{`} and \eqn{'} are removed;}
\item{doi entries are shown as URLs with labels in markdown grammar.}
}
}
\examples{
'stringi' |> citation() |> bibentry2text()
'texreg' |> citation() |> bibentry2text()
}
\keyword{internal}
