% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\name{c.md_lines}
\alias{c.md_lines}
\title{Combine Multiple \linkS4class{md_lines}}
\usage{
\method{c}{md_lines}(...)
}
\arguments{
\item{...}{one or more \linkS4class{md_lines} objects}
}
\value{
Function \code{\link[=c.md_lines]{c.md_lines()}} returns an \linkS4class{md_lines} object.
}
\description{
Combine Multiple \linkS4class{md_lines}
}
\keyword{internal}
