% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{citation_ext}
\alias{citation_ext}
\alias{url2doi}
\alias{dropManual}
\title{Handy Tools for \link[utils]{bibentry}}
\usage{
url2doi(x)

dropManual(x)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}
}
\value{
All functions returns a \link[utils]{bibentry} object
}
\description{
Handy Tools for \link[utils]{bibentry}
}
\details{
Function \code{\link[=url2doi]{url2doi()}} converts a \code{url} field to \code{doi}, if it is a DOI URL.

Function \code{\link[=dropManual]{dropManual()}} removes the CRAN \code{Manual} citation,
if a package has one-or-more \code{Article} citation.
}
\examples{
'stringi' |> citation() # using doi field, correct

'texreg' |> citation() # doi in url field, not good!
'texreg' |> citation() |> url2doi()
'survival' |> citation()
'survival' |> citation() |> dropManual()
}
\seealso{
\code{utils:::`[.bibentry`}
}
\keyword{internal}
