% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.R
\name{fromPackage}
\alias{fromPackage}
\title{Which Package Created This Object?}
\usage{
fromPackage(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=fromPackage]{fromPackage()}} returns a \link[base]{character} scalar.
}
\description{
Which Package Created This Object?
}
\note{
Function \code{\link[=fromPackage]{fromPackage()}} is not designed for use in package \pkg{rmd.tzh}, but for \pkg{ecip}.
}
\keyword{internal}
