% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_set.R
\name{key<-}
\alias{key<-}
\title{Set Key of \link[utils]{bibentry} and/or \link[utils]{citation}}
\usage{
key(x) <- value
}
\arguments{
\item{x}{a \link[utils]{bibentry} and/or \link[utils]{citation}}

\item{value}{\link[base]{character} scalar}
}
\value{
This syntactic sugar returns a \link[utils]{bibentry} and/or \link[utils]{citation} object.
}
\description{
Set Key of \link[utils]{bibentry} and/or \link[utils]{citation}
}
\keyword{internal}
