% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_pvalue_sym.R
\name{label_pvalue_sym}
\alias{label_pvalue_sym}
\title{Label \eqn{p}-values with Significance Symbol}
\usage{
label_pvalue_sym(...)
}
\arguments{
\item{...}{parameters of function \link[scales]{label_pvalue}}
}
\value{
Function \code{\link[=label_pvalue_sym]{label_pvalue_sym()}} returns a \link[base]{function}.
}
\description{
Label \eqn{p}-values with significance symbol.
}
\details{
Function \code{\link[=label_pvalue_sym]{label_pvalue_sym()}} is used so extensively,
therefore tzh keeps it in \pkg{rmd.tzh} instead of \pkg{scales.tzh}.

Pipeline:

Step 1: Apply function \link[scales]{label_pvalue}.

Step 2: Drop leading zeros.

Step 3: Apply function \link[stats]{symnum} (also see function \link[stats]{printCoefmat}).
}
\note{
Function \link[scales]{label_pvalue} is much prettier and more flexible than function \link[base]{format.pval}.
}
\examples{
p = c(a = pi^-100, b = .02, c = .05, d = .1, e = .9999, f = NA_real_)
p |> label_pvalue_sym()()
p |> label_pvalue_sym(add_p = TRUE)()

# below: exception handling
double() |> scales::label_pvalue()() # do not like!
# ?scales::pvalue is bad; ?scales::number is fine!
double() |> label_pvalue_sym()() # nice!
}
\keyword{internal}
