% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{md_.bibentry}
\alias{md_.bibentry}
\title{Markdown Lines of \link[utils]{bibentry} Object}
\usage{
\method{md_}{bibentry}(x, ...)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}

\item{...}{additional parameters, currently of no use}
}
\description{
Markdown Lines of \link[utils]{bibentry} Object
}
\examples{
list(
 stringi = 'String manipulation by <u>**`R`**</u> package <u>**`stringi`**</u>.' |> 
   new(Class = 'md_lines', package = 'stringi'),
 texreg = 'R regression output to LaTeX or HTML by <u>**`R`**</u> package <u>**`texreg`**</u>.' |> 
   new(Class = 'md_lines', package = 'texreg'),
 stats = 'Linear regression by <u>**`R`**</u> package <u>**`stats`**</u>.' |> 
   new(Class = 'md_lines', package = 'stats')
) |> render_(file = 'bibentry')

}
\keyword{internal}
