% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.binTab.R
\name{md_.binTab}
\alias{md_.binTab}
\title{R Markdown Lines for \link[DanielBiostatistics10th]{binTab}}
\usage{
\method{md_}{binTab}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[DanielBiostatistics10th]{binTab}
}
\note{
Do not want to \code{Imports: rmd.tzh} in package \CRANpkg{DanielBiostatistics10th}.
}
\examples{
library(DanielBiostatistics10th); list(
  '`binTab`' = array(c(7L, 3L, 8L, 6L), dim = c(2,2)) |> binTab()
) |> render_(file = 'binTab')
}
\keyword{internal}
