% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.htest.R
\name{md_.htest}
\alias{md_.htest}
\title{\link{md_.htest}}
\usage{
\method{md_}{htest}(x, xnm, ...)
}
\arguments{
\item{x}{an \code{htest}}

\item{xnm}{..}

\item{...}{..}
}
\value{
Function \code{\link[=md_.htest]{md_.htest()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
\link{md_.htest}
}
\keyword{internal}
