% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\docType{class}
\name{md_lines-class}
\alias{md_lines-class}
\title{\linkS4class{md_lines}}
\description{
..
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\link[base]{character} scalar or \link[base]{vector}}

\item{\code{bibentry}}{a \link[utils]{bibentry} object}
}}

\keyword{internal}
