% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.stats.R
\name{md_stats}
\alias{md_stats}
\alias{md_.factanal}
\title{Markdown Script for Objects from Package \pkg{stats}}
\usage{
\method{md_}{factanal}(x, ...)
}
\arguments{
\item{x}{an object returned by a function from package \pkg{stats}}

\item{...}{..}
}
\description{
Markdown Script for Objects from Package \pkg{stats}
}
\examples{
list(factanal = datasets::USJudgeRatings[-1L] |> stats::factanal(factors = 3L)) |>
  render_(file = 'factanal')

}
\keyword{internal}
