% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.R
\name{render_}
\alias{render_}
\title{Create and Render R Markdown file}
\usage{
render_(
  x,
  path = tempdir(),
  document = c("html", "word", "pdf"),
  file = stop("must specify `file` explicitly"),
  rmd.rm = TRUE,
  bib.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{an R object}

\item{path}{\link[base]{character} scalar}

\item{document}{\link[base]{character} scalar, type of output document,
\code{'html'} (default), \code{'word'} or \code{'pdf'}}

\item{file}{\link[base]{character} scalar}

\item{rmd.rm}{\link[base]{logical} scalar, whether to remove the R markdown \code{'.rmd'} file,
default \code{TRUE}}

\item{bib.rm}{\link[base]{logical} scalar, whether to remove the bibliography \code{'.bib'} file,
default \code{TRUE}}

\item{...}{..}
}
\description{
..
}
