% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{rmd.tzh-package}
\alias{rmd.tzh}
\alias{rmd.tzh-package}
\title{rmd.tzh: Markdown Lines for Common R Objects}
\description{
Markdown lines for common R objects.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
