% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{sort_by.bibentry}
\alias{sort_by.bibentry}
\title{Sort \link[utils]{bibentry} By}
\usage{
\method{sort_by}{bibentry}(x, y = "year", ...)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}

\item{y}{\link[base]{character} scalar}

\item{...}{additional parameters of function \link[base]{order}}
}
\description{
Sort \link[utils]{bibentry} By
}
\details{
Function \code{\link[=sort_by.bibentry]{sort_by.bibentry()}} sorts multiple citations of one package by some criteria (default being \code{'year'}).
}
\examples{
'rmarkdown' |> citation()
'rmarkdown' |> citation() |> sort_by(y = 'year', decreasing = TRUE)
}
