% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry.R
\name{subset.bibentry}
\alias{subset.bibentry}
\title{Subset \link[utils]{bibentry}}
\usage{
\method{subset}{bibentry}(x, subset, ...)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}

\item{subset}{..}

\item{...}{..}
}
\description{
Subset \link[utils]{bibentry}
}
\details{
Function \code{\link[=subset.bibentry]{subset.bibentry()}} ..
}
\examples{
(bib_survival = 'survival' |> citation())
bib_survival |> subset(subset = (bibtype != 'Manual'))
bib_survival |> subset(subset = (bibtype == 'Manual'))

(bib_ggplot2 = 'ggplot2' |> citation())
bib_ggplot2 |> subset(subset = (bibtype != 'Manual'))
bib_ggplot2 |> subset(subset = (bibtype == 'Manual'))
bib_ggplot2 |> subset(subset = (bibtype == 'Book'))
}
\keyword{internal}
