% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry.R
\name{subset_bibentry}
\alias{subset_bibentry}
\alias{subset.bibentry}
\alias{subset.citation}
\title{Subset \link[utils]{bibentry} and/or \link[utils]{citation}}
\usage{
\method{subset}{bibentry}(x, subset, ...)

\method{subset}{citation}(x, ...)
}
\arguments{
\item{x}{a \link[utils]{bibentry} or \link[utils]{citation} object}

\item{subset}{..}

\item{...}{..}
}
\description{
Subset \link[utils]{bibentry} and/or \link[utils]{citation}
}
\details{
Function \code{\link[=subset.bibentry]{subset.bibentry()}} ..

Function \code{\link[=subset.citation]{subset.citation()}} performs \code{\link[=subset.bibentry]{subset.bibentry()}} first,
then retains \code{attr(., which = 'package')}.
}
\examples{
(bib_survival = 'survival' |> citation())
bib_survival |> subset(subset = (bibtype != 'Manual'))
bib_survival |> subset(subset = (bibtype == 'Manual'))

(bib_ggplot2 = 'ggplot2' |> citation())
bib_ggplot2 |> subset(subset = (bibtype != 'Manual'))
bib_ggplot2 |> subset(subset = (bibtype == 'Manual'))
bib_ggplot2 |> subset(subset = (bibtype == 'Book'))

}
\keyword{internal}
