% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{url2doi}
\alias{url2doi}
\title{Prioritize \code{doi} over \code{url} in \link[utils]{bibentry}}
\usage{
url2doi(x)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}
}
\value{
Function \code{\link[=url2doi]{url2doi()}} returns a \link[utils]{bibentry} object.
}
\description{
Prioritize \code{doi} over \code{url} in \link[utils]{bibentry}
}
\details{
Function \code{\link[=url2doi]{url2doi()}} converts a \code{url} field to \code{doi} field in \link[utils]{bibentry},
if the \code{url} field is a DOI URL.
}
\examples{
'scales' |> citation() # using doi field, correct

'texreg' |> citation() # doi in url field, not good!
'texreg' |> citation() |> url2doi()
}
\keyword{internal}
