% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.md_lines.R
\name{c.md_lines}
\alias{c.md_lines}
\title{\code{\link[=c.md_lines]{c.md_lines()}}}
\usage{
\method{c}{md_lines}(...)
}
\arguments{
\item{...}{one or more \linkS4class{md_lines} objects}
}
\description{
\code{\link[=c.md_lines]{c.md_lines()}}
}
