% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.bibentry.R
\name{citation_ext}
\alias{citation_ext}
\alias{url2doi}
\alias{dropManual}
\alias{sortYear}
\title{Handy Tools for \link[utils]{citation}}
\usage{
url2doi(x)

dropManual(x)

sortYear(x, decreasing = TRUE)
}
\arguments{
\item{x}{a \link[utils]{citation} and/or \link[utils]{bibentry}}

\item{decreasing}{\link[base]{logical} scalar, see function \link[base]{order}}
}
\value{
All functions returns a \link[utils]{citation} and/or \link[utils]{bibentry}.
}
\description{
Handy Tools for \link[utils]{citation}
}
\details{
Function \code{\link[=url2doi]{url2doi()}} converts a \code{url} field to \code{doi}, if it is a DOI URL.

Function \code{\link[=dropManual]{dropManual()}} removes the CRAN \code{Manual} citation,
if a package has one-or-more \code{Article} citation.

Function \code{\link[=sortYear]{sortYear()}} sorts multiple citations of one package by year.
Tingting Zhan doesn't want to define \code{sort_by.citation()}, which sounds a little creepy :))
}
\examples{
'stringi' |> citation() # using doi field, correct
'texreg' |> citation() # doi in url field, not good!

# \pkg{texreg} has textVersion hard-coded in CITATION file
# therefore 'textVersion' cannot be updated
'texreg' |> citation() |> url2doi() # only a partial fix 

'robslopes' |> citation() |> toBibtex()
'robslopes' |> citation() |> url2doi() |> toBibtex() # complete fix!
'kernlab' |> citation()
'kernlab' |> citation() |> dropManual()
'lava' |> citation()
'lava' |> citation() |> sortYear()
}
\seealso{
\code{utils:::`[.bibentry`}
}
\keyword{internal}
