% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.DemographicTable.R
\name{md_.DemographicTable}
\alias{md_.DemographicTable}
\title{R Markdown Lines for \link[DemographicTable]{DemographicTable}}
\usage{
\method{md_}{DemographicTable}(x, xnm, ...)
}
\arguments{
\item{x}{a \link[DemographicTable]{DemographicTable}}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Lines for \link[DemographicTable]{DemographicTable}
}
\note{
Do not want to \code{Imports: rmd.tzh} in package \CRANpkg{DemographicTable}.
}
\examples{
library(DemographicTable); list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render_(file = 'DemographicTable')
}
\keyword{internal}
