% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.R
\name{pkg_text}
\alias{pkg_text}
\title{Text for Package to Create an R Object}
\usage{
pkg_text(x)
}
\arguments{
\item{x}{\link[base]{character} scalar, the returned object of function \code{\link[=fromPackage]{fromPackage()}}}
}
\value{
Function \code{\link[=pkg_text]{pkg_text()}} returns a \link[base]{character} scalar.
}
\description{
Text for Package to Create an R Object
}
\note{
Function \code{\link[=pkg_text]{pkg_text()}} is not designed for use in package \pkg{rmd.tzh}, but for \pkg{ecip}.
}
\keyword{internal}
