% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pkg_name.R
\name{pkg_text}
\alias{pkg_text}
\title{Text for Package to Create an R Object}
\usage{
pkg_text(x)
}
\arguments{
\item{x}{an R object}
}
\value{
Function \code{\link[=pkg_text]{pkg_text()}} returns a \link[base]{character} scalar.
}
\description{
Text for Package to Create an R Object
}
\keyword{internal}
