% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\alias{r_yaml_}
\title{YAML Header}
\usage{
r_yaml_(title, author = "tingting.zhan@jefferson.edu", document, bib, ...)
}
\arguments{
\item{title}{\link[base]{character} scalar}

\item{author}{\link[base]{character} scalar, author's name and/or email}

\item{document}{\link[base]{character} scalar}

\item{bib}{(optional) \link[base]{character} scalar, name of \code{.bib} bibliography file}

\item{...}{additional parameters, currently not in use}
}
\description{
YAML Header
}
\keyword{internal}
